requirements.txt
numpy
matplotlib
pandas
scipy
Pillow
psychopy
tqdm
psychopy-sounddevice